+++ GISLibrary/GM_Surface.cpp
@@ -91,12 +91,12 @@ namespace GM
 
 	int Surface::getInteriorRingCount()
 	{
-		return patch.boundary.interior.size();
+		return (int)patch.boundary.interior.size();
 	}
 
 	GM::Ring* Surface::getInteriorRing(int n)
 	{
-		if (n < 0 || n >= patch.boundary.interior.size())
+		if (n < 0 || n >= (int)patch.boundary.interior.size())
 		{
 			return nullptr;
 		}
